% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_display_plot.R
\name{display_plot}
\alias{display_plot}
\title{To display the plot with the certain annotations}
\usage{
display_plot(p, annots)
}
\value{
Return plot with the applied certain annotations
}
\description{
A utils function
}
